+++ js/app.js
@@ -399,6 +399,7 @@ const App = (() => {
     if (typeof LcPolish !== 'undefined') LcPolish.init();
     if (typeof LivePriceStream !== 'undefined') LivePriceStream.init();
     if (typeof PriceSnapshotService !== 'undefined') PriceSnapshotService.init();
+    if (typeof FundNavService !== 'undefined') FundNavService.applyAll();
     if (typeof GlanceBridge !== 'undefined') GlanceBridge.publish();
     _updateCurrencyToggleBtn();
     _scheduleAutoRefresh();